requirements.txt
Django==5.2.8
dj-database-url==3.0.1
python-dotenv==1.2.1
django-allauth==65.13.1
django-redis==6.0.0
redis==7.1.0
requests==2.32.5
pillow==12.0.0
whitenoise==6.11.0
celery==5.6.0
channels==4.3.2
channels-redis==4.3.0
daphne==4.2.1
psycopg2-binary==2.9.11
django-ratelimit==4.1.0
